feat(UserService, MusicService): relação n/m transformada para implicita
