+++ prisma/schema.prisma
@@ -17,7 +17,7 @@ model User {
   photo     String?
   password  String
   role      String
-  user      UserMusic[]
+  user      Music[]
 }
 
 model Artist {
@@ -35,14 +35,5 @@ model Music {
   album     String
   author    Artist  @relation(fields: [authorId], references: [id])
   authorId  Int   
-  music     UserMusic[] 
-  }
-
-model UserMusic {
-  user      User @relation(fields: [userId], references: [id])
-  userId    Int
-  music     Music @relation(fields: [musicId], references: [id])
-  musicId   Int
-
-  @@id([userId, musicId])
-}
+  music     User[] 
+  }
