+++ src/domains/music/services/MusicServices.ts
@@ -46,6 +46,34 @@ class MusicService {
         });
         return music;
     }
+
+    //Lista quais usuários já escutaram determinada música
+    async userWhoListenedMusic(musicId: number){
+
+    }
+
+    //Adiciona música a um usuário
+    async addMusicToUser(userId: number, musicId: number){
+
+    }
+
+    //Remove uma música de um usuário 
+    async removeMusicFromUser(userId: number, musicId: number){
+
+    }
+
+    //Verifica se um usuário já escutou determinada música
+    async hasUserHeardMusic(userId: number, musicId: number): Promise<boolean> {
+        const usersMusic = await prisma.userMusic.findUnique({
+                where: {
+                    userId_musicId: {
+                        userId: userId,
+                        musicId: musicId
+                    }
+                }
+            });
+        return usersMusic !== null;
+    }
 }
 
 export default MusicService;
